requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==1.10.13
python-dotenv==1.0.1
jinja2==3.1.3
python-multipart==0.0.6

# Security & Rate Limiting
slowapi==0.1.9
passlib[bcrypt]==1.7.4

# Logging & Monitoring
loguru==0.7.2
sentry-sdk[fastapi]==1.39.2

# Database Migrations
alembic==1.13.1

# Export & Reports
pandas==2.1.4
openpyxl==3.1.2
reportlab==4.0.8
matplotlib==3.8.2

# Caching
redis==5.0.1
hiredis==2.3.2
